requirements/requirements-gpu.txt
opencv-python==3.2.0.6
h5py==2.6.0
matplotlib==2.0.0
numpy==1.12.0
scipy==0.18.1
tqdm==4.11.2
keras==2.0.2
scikit-learn==0.18.1
pillow==4.0.0
ipykernel==4.6.1
tensorflow-gpu==1.0.0
django==1.11.9
djangorestframework==3.7.7
django-cors-headers==2.1.0